[NodeVisitor] make ReplaceDeprecatedConstant abstract
