+++ src/DependencyInjection/ContainerFactory.php
@@ -13,4 +13,12 @@ final class ContainerFactory
 
         return $appKernel->getContainer();
     }
+
+    public function createWithConfig(string $config): ContainerInterface
+    {
+        $appKernel = new AppKernel($config);
+        $appKernel->boot();
+
+        return $appKernel->getContainer();
+    }
 }
