+++ src/config/services.yml
@@ -5,7 +5,6 @@ parameters:
 services:
     _defaults:
         autowire: true
-        autoconfigure: true
 
     # PSR-4 autodiscovery
     Rector\:
