+++ tests/config/services.yml
@@ -1,0 +1,6 @@
+services:
+    _defaults:
+        autowire: true
+
+    # tests
+    Rector\Tests\NodeVisitor\UpgradeDeprecation\ReplaceDeprecatedConstantNodeVisitor\ReplaceOldConstantNodeVisitor: ~
